+++ src/libged/wdb_bigE.c
@@ -92,11 +92,11 @@ struct half_specific {
  */
 union E_tree
 {
-    long magic;
+    uint32_t magic;
 
     struct E_node {
 	/* the operator nodes */
-	long magic;
+	uint32_t magic;
 	int op;
 	union E_tree *left;
 	union E_tree *right;
@@ -104,7 +104,7 @@ union E_tree
 
     struct E_leaf {
 	/* the leaf nodes */
-	long magic;
+	uint32_t magic;
 	int op;
 	struct model *m;		/* NMG version of this leaf solid */
 	struct bu_list seghead;		/* head of list of segments for this leaf solid */
