+++ src/librt/primitives/nmg/nmg_visit.c
@@ -57,10 +57,10 @@ nmg_visit_vertex(struct vertex *v, const struct nmg_visit_handlers *htab, genptr
 {
     NMG_CK_VERTEX(v);
 
-    if (htab->vis_vertex) htab->vis_vertex((long *)v, state, 0);
+    if (htab->vis_vertex) htab->vis_vertex((uint32_t *)v, state, 0);
 
     if (htab->vis_vertex_g && v->vg_p)
-	htab->vis_vertex_g((long *)v->vg_p, state, 0);
+	htab->vis_vertex_g((uint32_t *)v->vg_p, state, 0);
 }
 
 
@@ -75,14 +75,14 @@ nmg_visit_vertexuse(struct vertexuse *vu, const struct nmg_visit_handlers *htab,
 {
     NMG_CK_VERTEXUSE(vu);
 
-    if (htab->bef_vertexuse) htab->bef_vertexuse((long *)vu, state, 0);
+    if (htab->bef_vertexuse) htab->bef_vertexuse((uint32_t *)vu, state, 0);
 
     nmg_visit_vertex(vu->v_p, htab, state);
 
     if (htab->vis_vertexuse_a && vu->a.magic_p)
-	htab->vis_vertexuse_a((long *)vu->a.magic_p, state, 0);
+	htab->vis_vertexuse_a((uint32_t *)vu->a.magic_p, state, 0);
 
-    if (htab->aft_vertexuse) htab->aft_vertexuse((long *)vu, state, 1);
+    if (htab->aft_vertexuse) htab->aft_vertexuse((uint32_t *)vu, state, 1);
 }
 
 
@@ -97,7 +97,7 @@ nmg_visit_edge(struct edge *e, const struct nmg_visit_handlers *htab, genptr_t s
 {
     NMG_CK_EDGE(e);
 
-    if (htab->vis_edge) htab->vis_edge((long *)e, state, 0);
+    if (htab->vis_edge) htab->vis_edge((uint32_t *)e, state, 0);
 }
 
 
@@ -112,15 +112,15 @@ nmg_visit_edgeuse(struct edgeuse *eu, const struct nmg_visit_handlers *htab, gen
 {
     NMG_CK_EDGEUSE(eu);
 
-    if (htab->bef_edgeuse) htab->bef_edgeuse((long *)eu, state, 0);
+    if (htab->bef_edgeuse) htab->bef_edgeuse((uint32_t *)eu, state, 0);
 
     nmg_visit_vertexuse(eu->vu_p, htab, state);
     nmg_visit_edge(eu->e_p, htab, state);
 
     if (htab->vis_edge_g && eu->g.magic_p)
-	htab->vis_edge_g((long *)eu->g.magic_p, state, 0);
+	htab->vis_edge_g((uint32_t *)eu->g.magic_p, state, 0);
 
-    if (htab->aft_edgeuse) htab->aft_edgeuse((long *)eu, state, 1);
+    if (htab->aft_edgeuse) htab->aft_edgeuse((uint32_t *)eu, state, 1);
 }
 
 
@@ -135,10 +135,10 @@ nmg_visit_loop(struct loop *l, const struct nmg_visit_handlers *htab, genptr_t s
 {
     NMG_CK_LOOP(l);
 
-    if (htab->vis_loop) htab->vis_loop((long *)l, state, 0);
+    if (htab->vis_loop) htab->vis_loop((uint32_t *)l, state, 0);
 
     if (htab->vis_loop_g && l->lg_p)
-	htab->vis_loop_g((long *)l->lg_p, state, 0);
+	htab->vis_loop_g((uint32_t *)l->lg_p, state, 0);
 }
 
 
@@ -153,7 +153,7 @@ nmg_visit_loopuse(struct loopuse *lu, const struct nmg_visit_handlers *htab, gen
 {
     NMG_CK_LOOPUSE(lu);
 
-    if (htab->bef_loopuse) htab->bef_loopuse((long *)lu, state, 0);
+    if (htab->bef_loopuse) htab->bef_loopuse((uint32_t *)lu, state, 0);
 
     if (BU_LIST_FIRST_MAGIC(&lu->down_hd) == NMG_VERTEXUSE_MAGIC) {
 	struct vertexuse *vu;
@@ -167,7 +167,7 @@ nmg_visit_loopuse(struct loopuse *lu, const struct nmg_visit_handlers *htab, gen
     }
     nmg_visit_loop(lu->l_p, htab, state);
 
-    if (htab->aft_loopuse) htab->aft_loopuse((long *)lu, state, 1);
+    if (htab->aft_loopuse) htab->aft_loopuse((uint32_t *)lu, state, 1);
 }
 
 
@@ -181,10 +181,10 @@ nmg_visit_face(struct face *f, const struct nmg_visit_handlers *htab, genptr_t s
     /* Handler's private state */
 {
 
-    if (htab->vis_face) htab->vis_face((long *)f, state, 0);
+    if (htab->vis_face) htab->vis_face((uint32_t *)f, state, 0);
 
     if (htab->vis_face_g && f->g.plane_p)
-	htab->vis_face_g((long *)f->g.plane_p, state, 0);
+	htab->vis_face_g((uint32_t *)f->g.plane_p, state, 0);
 }
 
 
@@ -201,7 +201,7 @@ nmg_visit_faceuse(struct faceuse *fu, const struct nmg_visit_handlers *htab, gen
 
     NMG_CK_FACEUSE(fu);
 
-    if (htab->bef_faceuse) htab->bef_faceuse((long *)fu, state, 0);
+    if (htab->bef_faceuse) htab->bef_faceuse((uint32_t *)fu, state, 0);
 
     for (BU_LIST_FOR(lu, loopuse, &fu->lu_hd)) {
 	nmg_visit_loopuse(lu, htab, state);
@@ -209,7 +209,7 @@ nmg_visit_faceuse(struct faceuse *fu, const struct nmg_visit_handlers *htab, gen
 
     nmg_visit_face(fu->f_p, htab, state);
 
-    if (htab->aft_faceuse) htab->aft_faceuse((long *)fu, state, 1);
+    if (htab->aft_faceuse) htab->aft_faceuse((uint32_t *)fu, state, 1);
 }
 
 
@@ -228,7 +228,7 @@ nmg_visit_shell(struct shell *s, const struct nmg_visit_handlers *htab, genptr_t
 
     NMG_CK_SHELL(s);
 
-    if (htab->bef_shell) htab->bef_shell((long *)s, state, 0);
+    if (htab->bef_shell) htab->bef_shell((uint32_t *)s, state, 0);
 
     for (BU_LIST_FOR(fu, faceuse, &s->fu_hd)) {
 	nmg_visit_faceuse(fu, htab, state);
@@ -241,9 +241,9 @@ nmg_visit_shell(struct shell *s, const struct nmg_visit_handlers *htab, genptr_t
     }
     if (s->vu_p) nmg_visit_vertexuse(s->vu_p, htab, state);
     if (htab->vis_shell_a && s->sa_p)
-	htab->vis_shell_a((long *)s->sa_p, state, 0);
+	htab->vis_shell_a((uint32_t *)s->sa_p, state, 0);
 
-    if (htab->aft_shell) htab->aft_shell((long *)s, state, 1);
+    if (htab->aft_shell) htab->aft_shell((uint32_t *)s, state, 1);
 }
 
 
@@ -260,15 +260,15 @@ nmg_visit_region(struct nmgregion *r, const struct nmg_visit_handlers *htab, gen
 
     NMG_CK_REGION(r);
 
-    if (htab->bef_region) htab->bef_region((long *)r, state, 0);
+    if (htab->bef_region) htab->bef_region((uint32_t *)r, state, 0);
 
     for (BU_LIST_FOR(s, shell, &r->s_hd)) {
 	nmg_visit_shell(s, htab, state);
     }
     if (htab->vis_region_a && r->ra_p)
-	htab->vis_region_a((long *)r->ra_p, state, 0);
+	htab->vis_region_a((uint32_t *)r->ra_p, state, 0);
 
-    if (htab->aft_region) htab->aft_region((long *)r, state, 1);
+    if (htab->aft_region) htab->aft_region((uint32_t *)r, state, 1);
 }
 /**
  * N M G _ V I S I T _ M O D E L
@@ -283,13 +283,13 @@ nmg_visit_model(struct model *model, const struct nmg_visit_handlers *htab, genp
 
     NMG_CK_MODEL(model);
 
-    if (htab->bef_model) htab->bef_model((long *)model, state, 0);
+    if (htab->bef_model) htab->bef_model((uint32_t *)model, state, 0);
 
     for (BU_LIST_FOR(r, nmgregion, &model->r_hd)) {
 	nmg_visit_region(r, htab, state);
     }
 
-    if (htab->aft_model) htab->aft_model((long *)model, state, 1);
+    if (htab->aft_model) htab->aft_model((uint32_t *)model, state, 1);
 }
 
 
@@ -297,7 +297,7 @@ nmg_visit_model(struct model *model, const struct nmg_visit_handlers *htab, genp
  * N M G _ V I S I T
  */
 void
-nmg_visit(const unsigned long *magicp, const struct nmg_visit_handlers *htab, genptr_t state)
+nmg_visit(const uint32_t *magicp, const struct nmg_visit_handlers *htab, genptr_t state)
     /* Handler's private state */
 {
     switch (*magicp) {
